+++ activerecord/lib/active_record/enum.rb
@@ -167,15 +167,6 @@ module ActiveRecord
       base.class_attribute(:defined_enums, instance_writer: false, default: {})
     end
 
-    def load_schema! # :nodoc:
-      attributes_to_define_after_schema_loads.each do |name, (cast_type, _default)|
-        unless columns_hash.key?(name)
-          cast_type = cast_type[type_for_attribute(name)] if Proc === cast_type
-          raise "Unknown enum attribute '#{name}' for #{self.name}" if Enum::EnumType === cast_type
-        end
-      end
-    end
-
     class EnumType < Type::Value # :nodoc:
       delegate :type, to: :subtype
 
@@ -255,6 +246,12 @@ module ActiveRecord
         detect_enum_conflict!(name, "#{name}=")
 
         attribute(name, **options) do |subtype|
+          if subtype == ActiveModel::Type.default_value
+            raise "Undeclared attribute type for enum '#{name}'. Enums must be" \
+              " backed by a database column or declared with an explicit type" \
+              " via `attribute`."
+          end
+
           subtype = subtype.subtype if EnumType === subtype
           EnumType.new(name, enum_values, subtype, raise_on_invalid_values: !validate)
         end